+++ tests/yaml/test_json.py
@@ -29,6 +29,25 @@ def test_TypeError_for_unhandled_types() -> None:
         dumps(object())
 
 
+def test_literal_string_can_be_dumped() -> None:
+    assert dumps("test") == '"test"'
+
+
+def test_callable_class_with_just_method_can_be_dumped() -> None:
+    class A:
+        def __call__(self) -> None:
+            pass
+
+    assert dumps(A()) == '"<tests.yaml.test_json.A()>"'
+
+
+def test_inline_function_can_be_dumped() -> None:
+    def function() -> None:
+        pass
+
+    assert dumps(function) == '"<tests.yaml.test_json.function>"'
+
+
 def test_LazyLoadConfiguration_can_be_dumped() -> None:
     config = LazyLoadConfiguration(ASSET_DIR / "config.yaml")
     typed = LazyLoadConfiguration(ASSET_DIR / "config.yaml").as_typed(Configuration)