+++ granular_configuration_language/__init__.py
@@ -1,12 +1,12 @@
 # isort:skip_file
-import granular_configuration_language.yaml.classes  # Needs to import before _configuration to prevent circular import
+# Order Matters
+from granular_configuration_language.yaml import Masked, Placeholder
 from granular_configuration_language._configuration import Configuration, MutableConfiguration
+import granular_configuration_language.proxy
 from granular_configuration_language._lazy_load_configuration import (
     LazyLoadConfiguration,
     LazyLoadConfiguration as LLC,
-    MutableLazyLoadConfiguration,
 )
-from granular_configuration_language._merge import merge  # depends on LazyLoadConfiguration
-
+from granular_configuration_language._mutable_lazy_load_configuration import MutableLazyLoadConfiguration
+from granular_configuration_language._merge import merge
 from granular_configuration_language._json import json_default
-from granular_configuration_language.yaml import Masked, Placeholder