+++ granular_configuration_language/_lazy_load_configuration.py
@@ -1,19 +1,19 @@
 from __future__ import annotations
 
 import collections.abc as tabc
-import copy
 import os
 import sys
 import typing as typ
-from collections.abc import Mapping, MutableMapping
+from collections.abc import Mapping
 from functools import cached_property
 from itertools import chain
 
+from granular_configuration_language import Configuration
 from granular_configuration_language._cache import NoteOfIntentToRead, prepare_to_load_configuration
-from granular_configuration_language._configuration import C, Configuration, MutableConfiguration
+from granular_configuration_language._configuration import C
 from granular_configuration_language._locations import Locations, PathOrStr
-from granular_configuration_language._simple_future import SimpleFuture
 from granular_configuration_language.exceptions import ErrorWhileLoadingConfig
+from granular_configuration_language.proxy import EagerIOConfigurationProxy, SafeConfigurationProxy
 
 if sys.version_info >= (3, 12):
     from typing import override
@@ -34,139 +34,6 @@ def _read_locations(
         env_locs = os.environ[env_location_var_name].split(",")
         load_order_location = chain(load_order_location, env_locs)
     return Locations(load_order_location)
-
-
-@Configuration.register  # pyright: ignore
-class SafeConfigurationProxy(Mapping):
-    """
-    Wraps a :py:class:`.LazyLoadConfiguration` instance to proxy all method and
-    attribute calls to its :py:class:`.Configuration` instance.
-
-    Passes ``isinstance( ... , Configuration)`` checks, as this class is
-    :py:meth:`~abc.ABCMeta.register`-ed as a subclass of
-    :py:class:`.Configuration`.
-
-    .. admonition:: Implementation Reasoning
-
-        Wrapping :py:class:`.LazyLoadConfiguration` maintains all laziness
-        build into :py:class:`.LazyLoadConfiguration`, while exposing all
-        of :py:class:`.Configuration`
-
-        This class is used behind a :py:func:`typing.cast` in
-        :py:meth:`.LazyLoadConfiguration.as_typed`, so it is not exposed
-        explicitly.
-
-    :param LazyLoadConfiguration llc:
-        :py:class:`.LazyLoadConfiguration` instance to be wrapped
-    """
-
-    __slots__ = ("__llc",)
-
-    def __init__(self, llc: LazyLoadConfiguration) -> None:
-        self.__llc = llc
-
-    def __getattr__(self, name: str) -> typ.Any:
-        return getattr(self.__llc.config, name)
-
-    @override
-    def __getitem__(self, key: typ.Any) -> typ.Any:
-        return self.__llc.config[key]
-
-    @override
-    def __iter__(self) -> tabc.Iterator[typ.Any]:
-        return iter(self.__llc.config)
-
-    @override
-    def __len__(self) -> int:
-        return len(self.__llc.config)
-
-    @override
-    def __contains__(self, key: typ.Any) -> bool:
-        return key in self.__llc.config
-
-    def __deepcopy__(self, memo: dict[int, typ.Any]) -> Configuration:
-        return copy.deepcopy(self.__llc.config, memo=memo)
-
-    def __copy__(self) -> Configuration:
-        return copy.copy(self.__llc.config)
-
-    copy = __copy__
-
-    @override
-    def __repr__(self) -> str:
-        return repr(self.__llc.config)
-
-
-def _eagerio_load(llc: LazyLoadConfiguration) -> Configuration:
-    return llc.config
-
-
-@Configuration.register  # pyright: ignore
-class EagerIOConfigurationProxy(Mapping):
-    """
-    .. versionadded:: 2.3.0
-
-    Wraps a :py:class:`.LazyLoadConfiguration` instance to proxy all method and
-    attribute calls to its :py:class:`.Configuration` instance.
-
-    Passes ``isinstance( ... , Configuration)`` checks, as this class is
-    :py:meth:`~abc.ABCMeta.register`-ed as a subclass of
-    :py:class:`.Configuration`.
-
-    .. admonition:: Part of the EagerIO feature set
-        :class: caution
-
-        This immediately spawns a thread to load and build the Configuration in
-        the background, so that future calls are non-/minimally blocking.
-
-        This class is used behind a :py:func:`typing.cast` in
-        :py:meth:`.LazyLoadConfiguration.eager_load`, so it is not exposed
-        explicitly.
-
-    :param LazyLoadConfiguration llc:
-        :py:class:`.LazyLoadConfiguration` instance to be wrapped
-    """
-
-    def __init__(self, llc: LazyLoadConfiguration) -> None:
-        self.__future = SimpleFuture(_eagerio_load, llc)
-
-    @cached_property
-    def __config(self) -> Configuration:
-        try:
-            return self.__future.result
-        finally:
-            del self.__future
-
-    def __getattr__(self, name: str) -> typ.Any:
-        return getattr(self.__config, name)
-
-    @override
-    def __getitem__(self, key: typ.Any) -> typ.Any:
-        return self.__config[key]
-
-    @override
-    def __iter__(self) -> tabc.Iterator[typ.Any]:
-        return iter(self.__config)
-
-    @override
-    def __len__(self) -> int:
-        return len(self.__config)
-
-    @override
-    def __contains__(self, key: typ.Any) -> bool:
-        return key in self.__config
-
-    def __deepcopy__(self, memo: dict[int, typ.Any]) -> Configuration:
-        return copy.deepcopy(self.__config, memo=memo)
-
-    def __copy__(self) -> Configuration:
-        return copy.copy(self.__config)
-
-    copy = __copy__
-
-    @override
-    def __repr__(self) -> str:
-        return repr(self.__config)
 
 
 class LazyLoadConfiguration(Mapping):
@@ -448,126 +315,3 @@ class LazyLoadConfiguration(Mapping):
         :rtype: C
         """
         return typ.cast(C, EagerIOConfigurationProxy(self))
-
-
-class MutableLazyLoadConfiguration(LazyLoadConfiguration, MutableMapping):
-    r"""
-    Used to define a **mutable** Configuration from file paths
-    that lazily loads on first access.
-
-    **Options:**
-        -  Using ``env_location_var_name``, you can enable pulling locations
-           from an environment variable.
-
-    .. tip::
-
-        Consider using an immutable configuration with
-        :py:class:`.LazyLoadConfiguration` in you code to reduce unexpected
-        side-effects.
-
-    .. admonition:: Classes used for mutability
-        :class: note
-        :collapsible: closed
-
-        - :py:class:`.MutableConfiguration` for mappings
-        - :py:class:`list` for sequences
-
-    :param ~pathlib.Path | str | os.PathLike \*load_order_location:
-            File path to configuration file
-    :param str | ~collections.abc.Sequence[str], optional base_path:
-        Defines the subsection of the configuration file to use.
-        See Examples for usage options.
-    :param bool, optional use_env_location:
-        - Enabled to use the default environment variable location.
-        - Setting to :py:data:`True` is only required if you don't change
-          ``env_location_var_name`` from its default value.
-    :param str, optional env_location_var_name:
-        - Specify what environment variable to check for additional file paths.
-        - The Environment Variable is read as a comma-delimited list of
-          configuration path that will be appended to ``load_order_location``
-          list.
-        - Setting the Environment Variable is always optional.
-        - *Default*: ``G_CONFIG_LOCATION``
-
-          - Setting ``use_env_location=True`` is required to use the default
-            value.
-
-    :examples:
-        .. code-block:: python
-
-            # Base Path - Single Key
-            MutableLazyLoadConfiguration(..., base_path="base_path")
-            # Base Path - JSON Pointer (strings only)
-            MutableLazyLoadConfiguration(..., base_path="/base/path")
-            # Base Path - List of keys
-            MutableLazyLoadConfiguration(..., base_path=("base", "path"))
-
-            # Use Environment Variable: "CONFIG_LOC"
-            MutableLazyLoadConfiguration(..., env_location_var_name="CONFIG_LOC")
-
-            # Use default Environment Variable: "G_CONFIG_LOCATION"
-            MutableLazyLoadConfiguration(..., use_env_location=True)
-    """
-
-    def __init__(
-        self,
-        *load_order_location: PathOrStr,
-        base_path: str | tabc.Sequence[str] | None = None,
-        use_env_location: bool = False,
-        env_location_var_name: str = "G_CONFIG_LOCATION",
-    ) -> None:
-        super().__init__(
-            *load_order_location,
-            base_path=base_path,
-            use_env_location=use_env_location,
-            env_location_var_name=env_location_var_name,
-            inject_before=None,
-            inject_after=None,
-            disable_caching=True,
-            _mutable_configuration=True,
-        )
-
-    @property
-    @override
-    def config(self) -> MutableConfiguration:
-        """
-        Load and fetch the configuration. Configuration is cached for
-        subsequent calls.
-
-        .. admonition:: Thread-safe
-            :class: tip
-            :collapsible: closed
-
-            Loading the configuration is thread-safe and locks while the
-            configuration is loaded to prevent duplicative processing and data
-
-        """
-        return typ.cast(MutableConfiguration, super().config)
-
-    @override
-    def __delitem__(self, key: typ.Any) -> None:
-        del self.config[key]
-
-    @override
-    def __setitem__(self, key: typ.Any, value: typ.Any) -> None:
-        self.config[key] = value
-
-    @override
-    def as_typed(self, typed_base: type[C]) -> typ.NoReturn:
-        """
-        Not supported for :py:class:`MutableLazyLoadConfiguration`.
-        Use :py:class:`LazyLoadConfiguration`.
-        """
-        raise NotImplementedError(
-            "`as_typed` is not supported for `MutableLazyLoadConfiguration`. Use `LazyLoadConfiguration`."
-        )
-
-    @override
-    def eager_load(self, typed_base: type[C]) -> typ.NoReturn:
-        """
-        Not supported for :py:class:`MutableLazyLoadConfiguration`.
-        Use :py:class:`LazyLoadConfiguration`.
-        """
-        raise NotImplementedError(
-            "`eager_load` is not supported for `MutableLazyLoadConfiguration`. Use `LazyLoadConfiguration`."
-        )