+++ granular_configuration_language/proxy/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from granular_configuration_language._lazy_load_configuration import (  # noqa: F401
+from granular_configuration_language.proxy._proxy import (
     EagerIOConfigurationProxy,
     SafeConfigurationProxy,
 )